pyproject.toml
[project]
name = "french-towns-lakehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.4.4",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pandas>=3.0.0",
    "prefect>=3.6.17",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
